backend/requirements.txt
# MagiStock — Dependencies
# Install: pip install -r requirements.txt

# Agentfield SDK
agentfield

# AI & Structured Output
pydantic>=2.0

# Numerical Computing (used in Skills)
numpy>=1.24
pandas>=2.0

# Environment Variables
python-dotenv

# Market Data (optional — falls back to synthetic data if unavailable)
yfinance
